+++ src/lib/retirement/index.ts
@@ -1,2 +1,3 @@
 export * from "./constants";
+export * from "./drawdown";
 export * from "./project";